Added the whole selection range restoration upon popup dismissal
