+++ packages/text-annotator-react/src/hooks/index.ts
@@ -1,1 +1,1 @@
-export { useRestoreSelectionCaret } from './useRestoreSelectionCaret';
+export { useRestoreSelectionRange } from './useRestoreSelectionRange';
